Migrate vault_sorted_json to v4 function API
